requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2

# Data processing
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
bs4==0.0.1

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2
uuid==1.30

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
